Move the useGetLocation hook to base hooks
