+++ plugins/woocommerce-blocks/assets/js/base/hooks/index.js
@@ -1,4 +1,5 @@
 export * from './use-container-queries';
+export * from './use-get-location';
 export * from './use-local-storage-state';
 export * from './use-position-relative-to-viewport';
 export * from './use-previous';
